Провайдерю SizeCalculator. Сделал инжект презентера со скоупом активити(буду доделывть)
