+++ app/src/main/java/com/andersen/internship/filmsapp/ui/adapters/FilmItemAdapter.kt
@@ -9,14 +9,15 @@ import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
 import com.andersen.internship.filmsapp.R
+import com.andersen.internship.filmsapp.SizeCalculator
+import com.andersen.internship.filmsapp.SizeOfView
 import com.andersen.internship.filmsapp.pojo.films.Film
 import com.squareup.picasso.Picasso
 import kotlinx.android.synthetic.main.item_films.view.*
 import kotlin.math.roundToInt
 
-class FilmItemAdapter(private val activity: Activity) : RecyclerView.Adapter<FilmItemAdapter.FilmsHolder>() {
+class FilmItemAdapter(private val sizeOfImageView: SizeOfView) : RecyclerView.Adapter<FilmItemAdapter.FilmsHolder>() {
 
-    private lateinit var sizeOfImageView: SizeOfView
 
     var listFilms = emptyList<Film>()
         set(value) {
@@ -26,8 +27,7 @@ class FilmItemAdapter(private val activity: Activity) : RecyclerView.Adapter<Fil
 
     override fun onCreateViewHolder(viewGroup: ViewGroup, p1: Int): FilmsHolder {
 
-        sizeOfImageView = SizeCalculator(activity).calculateWidthAndHeightOfView()
-        val cardView = LayoutInflater.from(activity).inflate(R.layout.item_films, viewGroup, false)
+        val cardView = LayoutInflater.from(viewGroup.context).inflate(R.layout.item_films, viewGroup, false)
         return FilmsHolder(cardView)
     }
 
@@ -51,45 +51,5 @@ class FilmItemAdapter(private val activity: Activity) : RecyclerView.Adapter<Fil
         val imageViewPoster: ImageView = itemView.imageViewPoster
     }
 
-    class SizeCalculator(private val activity: Activity){
-
-        private var widthScreen: Int = 0
-
-
-        fun calculateWidthAndHeightOfView(): SizeOfView{
-
-            val recyclerViewMargin = activity.resources.getDimension(R.dimen.recycler_view_margin).roundToInt()
-            val linearLayoutPadding = activity.resources.getDimension(R.dimen.linear_layout_padding).roundToInt()
-            val cardViewLayoutMargin = activity.resources.getDimension(R.dimen.card_view_layout_margin).roundToInt()
-            val spanCount = calculateSpanCount()
-
-            val widthImage = (calculateWidthScreen() - 2 * (recyclerViewMargin - spanCount * (linearLayoutPadding - cardViewLayoutMargin)))/spanCount
-            val heightImage = widthImage * 3/2
-
-            return SizeOfView(widthImage, heightImage)
-        }
-
-        fun calculateSpanCount(): Int{
-
-            var spanCount: Int = calculateWidthScreen()/IMAGET_VIEW_APPROXIMATE_WIDTH
-            if(spanCount < MIN_SPAN_COUNT) spanCount = MIN_SPAN_COUNT
-            return spanCount
-        }
-
-        fun calculateWidthScreen():Int{
-            if (widthScreen == 0) {
-                val display = activity.windowManager.defaultDisplay
-                val size = Point()
-                display.getSize(size)
-                widthScreen = size.x
-            }
-            return widthScreen
-        }
-    }
-
-    data class SizeOfView(val width: Int, val height: Int)
 }
 
-private val IMAGET_VIEW_APPROXIMATE_WIDTH = 256
-private val MIN_SPAN_COUNT = 2
-private val HEIGHT_WIDTH_RELATION = 3/2
